+++ src/main/java/fr/estecka/variantscit/VariantsCitMod.java
@@ -1,8 +1,6 @@
 package fr.estecka.variantscit;
 
 import net.fabricmc.api.ClientModInitializer;
-import net.minecraft.client.render.item.model.ItemModelTypes;
-import net.minecraft.client.render.item.property.numeric.NumericProperties;
 import net.minecraft.component.DataComponentTypes;
 import net.minecraft.item.Item;
 import net.minecraft.util.Identifier;
@@ -14,8 +12,6 @@ import org.slf4j.LoggerFactory;
 import com.mojang.serialization.MapCodec;
 import fr.estecka.variantscit.reload.ModuleLoader;
 import fr.estecka.variantscit.modules.*;
-import fr.estecka.variantscit.properties.*;
-import fr.estecka.variantscit.selectors.*;
 
 
 public class VariantsCitMod
@@ -56,13 +52,6 @@ implements ClientModInitializer
 		ModuleRegistry.Register(Identifier.ofVanilla("trim"), new TrimModule());
 		ModuleRegistry.Register(Identifier.ofVanilla("trim_pattern"), new TrimPatternModule());
 		ModuleRegistry.Register(Identifier.ofVanilla("trim_material"), new TrimPatternModule());
-
-		NumericProperties.ID_MAPPER.put(Identifier.of(MODID, "block_entity_data"), NbtNumberProperty.CreateCodec(DataComponentTypes.BLOCK_ENTITY_DATA));
-		NumericProperties.ID_MAPPER.put(Identifier.of(MODID, "bucket_entity_data"), NbtNumberProperty.CreateCodec(DataComponentTypes.BUCKET_ENTITY_DATA));
-		NumericProperties.ID_MAPPER.put(Identifier.of(MODID, "custom_data"), NbtNumberProperty.CreateCodec(DataComponentTypes.CUSTOM_DATA));
-		NumericProperties.ID_MAPPER.put(Identifier.of(MODID, "entity_data"), NbtNumberProperty.CreateCodec(DataComponentTypes.ENTITY_DATA));
-		NumericProperties.ID_MAPPER.put(Identifier.of(MODID, "stored_enchantment_level"), EnchantedBookLevelPredicate.CODEC);
-		ItemModelTypes.ID_MAPPER.put(Identifier.ofVanilla("range_dispatch"), DynamicRangeDispatchUnbaked.CODEC);
 	}
 
 	static public void OnResourceReload(ModuleLoader.Result result){