+++ src/main/java/fr/estecka/variantscit/format/properties/AxolotlVariantProperty.java
@@ -1,27 +1,19 @@
 package fr.estecka.variantscit.format.properties;
 
 import net.minecraft.component.DataComponentTypes;
-import net.minecraft.component.type.NbtComponent;
 import net.minecraft.entity.passive.AxolotlEntity;
-import net.minecraft.nbt.NbtCompound;
-import net.minecraft.nbt.NbtElement;
 
 public class AxolotlVariantProperty
-extends AMonoComponentProperty<NbtComponent>
+extends AMonoComponentProperty<AxolotlEntity.Variant>
 {
 	static public final AxolotlVariantProperty UNIT = new AxolotlVariantProperty();
 
 	public AxolotlVariantProperty(){
-		super(DataComponentTypes.BUCKET_ENTITY_DATA);
+		super(DataComponentTypes.AXOLOTL_VARIANT);
 	}
 
 	@Override
-	public String GetPropertyString(NbtComponent component) {
-		NbtCompound nbt;
-		if ((nbt=component.getNbt()) == null || !nbt.contains("Variant", NbtElement.NUMBER_TYPE))
-			return null;
-
-		int variantRaw = nbt.getInt("Variant");
-		return AxolotlEntity.Variant.byId(variantRaw).getName();
+	public String GetPropertyString(AxolotlEntity.Variant component) {
+		return component.getId();
 	}
 }