+++ src/main/java/fr/estecka/variantscit/modules/GoatHornModule.java
@@ -2,9 +2,8 @@ package fr.estecka.variantscit.modules;
 
 import fr.estecka.variantscit.api.ISimpleCitModule;
 import net.minecraft.component.DataComponentTypes;
-import net.minecraft.item.Instrument;
+import net.minecraft.component.type.InstrumentComponent;
 import net.minecraft.item.ItemStack;
-import net.minecraft.registry.entry.RegistryEntry;
 import net.minecraft.util.Identifier;
 
 public class GoatHornModule
@@ -12,10 +11,10 @@ implements ISimpleCitModule
 {
 	@Override
 	public Identifier GetItemVariant(ItemStack stack){
-		RegistryEntry<Instrument> component = stack.get(DataComponentTypes.INSTRUMENT);
+		InstrumentComponent component = stack.get(DataComponentTypes.INSTRUMENT);
 		if (component == null)
 			return null;
 
-		return component.getKey().get().getValue();
+		return component.instrument().getKey().get().getValue();
 	}
 }